requirements.txt
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
httplib2==0.31.2
idna==3.13
oauthlib==3.3.1
proto-plus==1.27.2
protobuf==7.34.1
psycopg==3.3.3
psycopg-binary==3.3.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
